12-18-22 Commit v2.5.3
12-18-22 v2.5.3
12-18-22 Commit
With the help of Paul in a forum, the physics hands now start rotated 90/-90 degrees inward so that the palms are facing inward (as opposed to palms facing towards the ground) on start.  This orients the hands in the same direction as the player's physical hands while holding the VR hand controllers.  This rotation is also stable and is no longer causing the hands to shake uncontrolably due to that rotation every frame, which was part of the solution Paul helped me with.

As part of this solution I renamed my HandPhysics.cs script to HandPhysicsPosRight and I duplicated this script and I renamed it to HandPhyisicsPosLeft.  The then assigned these to the proper left and right physics hands and I relinked the proper game object and mesh render to each.  The only difference between these two scripts is that one script has a rotation of 90 degrees and the other script has a rotation of -90 degrees.

I also rotated adjusted the starting rotation of the non-physics hands to match so that when the non-physics hands are shown, the rotation matches those of the physics hand model at the time of collision.

I fixed the reset button in the Peaceful Nowhere scene by changing the Clicker's threshold from 1.0f to 1.5f so that the button pressed a little further and registered as a click and release (thus triggering the script that resets the scene).

I fixed the spinning ghost hands at the start of a new scene and when restarting a scene using a button by changing the Y and X position transform of the physics hands and then matching the position transform of the non-physics controllers to match. The hands are now closer to the actual starting position that the physical controllers are likely to be and not on the floor at the start of the scene.

Lastly, I also updated these physics and non-physics hand changes in the Peaceful Nowhere scene by updating my -- XR -- prefab.  I then unpacked the -- XR -- prefab in my Hierarchy completely in my Matt's Room scene.  I then loaded the Peaceful Nowhere scene where I deleted -- XR -- from the Hierarchy pane and I brought back in -- XR -- from the prefabs.  I then unpacked the -- XR -- prefab in my Hierarchy completely in my Peaceful Nowhere scene.

I also made my Projectiles parent object a prefab so that my custom dart gun and stapler could be exported or duplicated, if need be.
@mattdway
mattdway committed on Dec 18